Generate an HTML widget to download input Rmd file

The button can self-contain the input data by base64 encoding.

download_rmd_button(
  input = NULL,
  text = "Download Rmd",
  ...,
  class = "button",
  align = "right",
  aside = TRUE,
  embed = xfun::embed_file
)

Arguments

input
Filename of the input. If NULL, the function automatically detects the name of the knitting Rmd file.

text
The text for the hyperlink.

...
Arguments passed to the embed function.

class
HTML class(es) given to the anchor element generated by embed (default: "button").

align
Align the button by CSS's text-align attribute. This feature is disabled when aside is FALSE

aside
Whether to wrap the anchor element by the aside element.

embed
A function to embed file(s). One of xfun::embed_file, xfun::embed_files, or xfun::embed_dir.

Value

shiny.tag class object.

Examples

set.seed(1L)

input <- tempfile()
writeLines("", input)
download_rmd_button(input)

#> <aside style="text-align:right">
#>   <a href="data:text/plain;base64,Cg==" download="file22ade8c5126" class="button">Download Rmd</a>
#> </aside>

input <- tempdir()
download_rmd_button(input, embed = xfun::embed_dir)

#> <aside style="text-align:right">
#>   <a href="data:application/zip;base64,UEsDBBQAAgAIAD1JzlAgd+Ci0DAAAAQxAAAfAAAAbGlibG9jXzE4MF9hMzhmN2Q1ODEzMzQ0YWY5LnJkc52aU28uAIOt6/atbdu2bdu2bdu2bdu2bbu72Lt2e75zMpnkJHM162IlT9YPWDcPQQwU0P8L6K3fjj2pRLpmzRT5VtW63nYa63Ik9u18ZAmxGsa+d6yWgkcLWdtqHcxWsnsvvk1LdXO1cm01rZl509Q4Nsyn1YZ/41EEeO/gsD+7oY9d7MFXq8IPCKj/C0g4sMAbzHr8DpeTvKlp06S5DLYstCUZCQ0zPcX8d27k28jhcpkMTAZ6+ugAzfGCjAb/wO/+x12ZTIYBYtjM8UHL8QFKBjqDHtqCwQDP3LJHhBQFbj9P7M7fu4GcMa9W+h/ZkafSF7FUcyfmer/scSnsC4q2PkvkuB6t23nL5FIv9R6cBIQTp19/ZZ7Ml08lcv09fIRhJ4rkAbdiJ/ShxxrOzz99CvDwjbnRf6x2p1oayovXKSANzihSKt1JbmGbt/v66my5q/1wxfZu1JmzR0cYM0C7uOIiujLDM0dcAszP/53CMY0jFDfSQNjo00VsHZHOFmJGCpSwLkAa2e/rwj1auS/L4cc2Rfs6kWZHFkrz3OBWa21jaLX2YcXxnI9J80Q3QxpxM5/R2N3vDKLSH3wBjXuwyHO/DjttJig/WhfhXbwp5To4ZA8nU5KiXDo28club24wXqXc17emnCamYuxYPQHlyKaAQZKlu0N969+GogAkcCedKXoLSEXAVCmQ00aa8K7W5gn/4MYP4w42L2OVTlEOPQhbYI5cjhUcmQbjIZxIRcMhnNyYrpbrkY2QzpEbUTQ7iJeY8kKviGSqUnA1k8jhy4ZLxfuVHUkxjyuVS6WfHf1DtLHQEojr2VxkxTPeS4kX9AxSXXTgmB2J47VFcJL34fgrKLD3YuoJEqJ5I45V78/BPU1183LcWBBeotSacDCeTNpIyp13ulml2WNEnIR+uEaGp7ffk1BwdvodrVngmTyGlZ4wfLRaP792yakwos8c7p8/iqEnWEbT55h7lP08/DRZLMIw+6DEOWSMeg765daEAXdUkabPGwuGCXyo8By+5MQxC3FLKgt8o3EmA715pr9MQHq1ib4iHwlGHgOyAuSb4fMS8dWeCyFvmF1O9Z9SULk0rCIo99vRG5DRsCwjAMwm8EgufvtzxgPsGV4/OUhxAtn+vOJBY6CdJkr5RSXL6ciCB+AaeyCDgkMM+n6A3nDi55KGejc8nfRGxSximBIcV2J8jlTl7VPG0rHNvFAZqn9Umc9ukhAbFNCPpAl5ohJdIMLE3iIUn7jHqrJFfoYUvUX0SvcL4rNjEhFCvltKEn2X+t4LyuNjeFglXqRcIblnX0SuhFxxohiCdtKNBceSA9UQ4r+CAYWKAvdqRk4E9s04kQuyQfjqvHPLjI9gbrBwpqaYV7pwxttNjDTiwYW9Bcf/DQ9RirQhFo03C2l23I2Np3jZiaBIfjMkHaFE3YRP9J5H5SC3ckZFmXJ1w3wlngpypBsFchSdCZY5X3qbRH/0taxy4eSSDJeQvB6x0iEMFDsoU9pMSCLWEJWatOXRmdIEzOv1YEDf+/gyXIl32s+UWKaOUoIU4I8FN1FXiXc9UOIXsWJGjRHG0U89XHg0cg0RFj+BBaT6HdmKG2/dVjMkyFNC6OEUuHpDvzaLw6SzXD2H3gl266w7NE6JWlJohctBHzlnFs8aIIv19NKgECVIqcCkIJJ7uiFDswUCqfdO8oRAAoINpShJRxC5XkUPEalkEdU0yCO+2xW1ukt3AWGSY+FiXFivx/zw74nSXlqpwiDjb1cSHo1lRpzzFSVE7jRPoKdRyTRwwDcxov60sog7GuRT/0SnLb8o4W6XIVNYpx9T4jWZCq2+zamhKFsKEE/1kDrpn/pvMCELyPmCFFs9TckyKKHds5YeJJ8SbB2Sta0iqZfSAEbfNAey4LDLUEEdA0JBn6c/45TTkFIcPECw2rsHksg6PtkTkY4Zrx1FWkN+5HuYUu5UaMJLzsGtLH+oy6AY3X7TV03tob9npKe/4f8IaZwddISRB1af+3dkkU9QvMytsULzpAoxVGEEis0Ur36b/gUyJIyhqtriw9/OJomq5kTRs00uCDs3p0CPYuNrT2sR4ozKbvdFDRtpKHoJruvOhC+vlCslCJeY1FcJSvPLTRtxGo2xZ5ekzUx/+ZuH6+uOwbMqaHX84Zdt2jWIpHNmTd14alO0V4xfYB7sP65EMaGG1hYFJtniux2X4oh8stlCIprJmgBOIY66u/Co75LM0XhPYvdL+BKFUblAV7o5b2xuF1cmS7m7edp86GjsOIkbihWMfZsCsmV2LTxI9vZJiF9w2NHs+AhUQ6FS5K1I7sUlIncreF/rw4hCplBnqpOevt5J8EhML5JegEhFcihieinsHqys95UlQVIle2wLzn5mOVM7idzUnP7QKdcaz3tHLAXZXBgEupjuUAncnhtzdRHpevw4dQoGs5Kf5vMeQol98hbGloWQuEAr/7hP/8MnlnI4naSJrrtZolkptzioRDtKzf6vlEq5Ftn25oSSAsToXWgVvTIukFFM8KymmdaMqPgKlGlYRAYKAUKKM/is64FwXmlOmwv+TLtoV5v3Tk7J/Puor+fNmzV0vz5qQCiUyAxfZgtqj6nM955UlKnZlxiVC6Y3LIISX7/giqqSWi2tWlqtrFpZrapaVa36f4XLDt5laePiJV1LuhGk0+BK7AH6VQ6sVkHfyh3g9/4mOao6DksQ2q3aIy/dK7uVJnNn4+IyvloUIdoOtnasdiIbGx62Is/PdjqNJrqocjfeP+cGqxzt7O2Fi9ubrSS2yUf0UUQ39qtXh9ft+ngwQWBhmPp9+4UCHFSbf9obRXJsl4fJ9YLxePrptKvkYYxv0H3huolbV8Lk+sHcmIfOx+s9go1Z3SFsc0EEfXmBpso+xl/EXWbiDXEiWhQngpqpJsmslwgm33Rq18nffSaEowbdXEKCBnq1mIfJ80r0ms2iCGE3OtjrVRjtlgz+uHPyoPrqKwNWhyVNrZLVNs9RqnwkNCskxYD/GmPa1JlNLBXTaBxslsk5LJXj4O+KlyKFNMLQSIOEtfeulMP0CXYGlp6mh7mmLJSKMc0V7wWDFwFpumQwwFe1yRx09CraTa6VhZU7fMHQyVeaLdqnOc5ADAVArIhx+fMY4pYYwk4CBUMM9l3lgYjAQu41PFrHWiQV3T1IAfkz2E05paDtCSUmi+k2sg6OhZ0ZgfsH3+jIfnR/D8NbawEw17EvimODNq4YYK8cOQJtdVVrOy5FAQM7TyZJ1IkG9S7tf8xi3IcwYeyUDjduecEjuSK2X9Bz0/aOH0sbm5qb5p3W4J28VGZDiS7pMbYXQ08ln5+Zs1YOwYePASVbp71hRtgou3IwA+GgAgAbYjZtHFssp5BExDNREDbMPB3cDNmQpvTrVH/2rlBB+ZVk0q2neKE/FJQMp4scTsYaa2+RhQa2mmPmXXfQ1uSWHVWwVktj1u9JurH9Rwxg0VnMfaLFp+V6jOQrmL8UX9/cQUB5ZcYL+/FnqODHRI0euPet8kTRYWRTydi1btVAiGuVYvosYFttBAvJ7zmthxO2U5PZpTl2BI7pZJx0WNHeO7FKinZAW8WDYXjDilvLtisgKo+YC0Lks5LDCPRl9UcbdPLTbgOLGnzJ49mlGdxkb3TbqCi+sNVg+3HPct20stdG0mkaSKrX3nnqIi9KiyqbsZ0/WsXGi27llW3EoSfPbuOReNqmubnNBDw75TaSTFd9VF3jvyi/agt/tPhqLJ5r89davPU3VyqlPmscfocWNMI+ms33dUM4hJ2c9qVwpHppHIYOIN+7AXT22/rYNm1r7rk1zQ/37GMwbrUidmu4P4o6c85e44vvF+wupnRVBfNMAY/xHfeks/VKDO4jNf2iOUNTaWrWjdLzgu3s8rd264UNS/KFJmDtkCAs/0JZhfpVYyo97nPSnoc6r7w2/snNvepNWXPAmNh7ryMdbD3vUPdv5zq4ujCIuuSHZTrs8t+sD/QKFanqrEXYHBGjhgv28MRx8TrIz1BP1BI441Y8Txk3oO80WW5kg0evad1yfBvkZz2zfRny59W396LSh2fdiUk08So5qeHFonmsGjh1tFsvBPxTAncXcDGOdWGFUzjLNYFQZuPQtldR0IKmPM74dpoRNuMv1aM3sphzzmb4LD0yowjLmjVdcGGV2m8qnx9yxT7rfBMOek1Xb1t08HRaJrmZp6Tnk4U+5yPwm0sWaTnd/KLLqLtX59pKf5QRcNgY+KZf8w3ND56I+JBolkbqPj9yrijJVWo2v892qb/OgjKEAEJqMQhDcmjAgqtrn0ZgKvYFy2lnAOh8+r0oLwbxvc1E1uPxpHJotjQ7J2dqyxHUjHIwIE/dOU8SmKJYmuRFlzg/V119zh3FW6hJWaxx5tWb/ys55k4/BoVOrtOivvRVrtYXTNdbu9yf46nWPdikJfjSyf1IlNpO0HmsubVQOuX1Dft9qsWzsUvL3TkFvauSYcnIE4qNMK+6P0zpIlOPkEHUKYvzXj3iyrySjXdDLzHGl2eVcc8yYLfQFCLzNqnafG36kwWt2Be4BNeFfqrJPHjxvhrH8ifCoqncM28Drz2WVHJBL7dKcI6VmK7lyfDrivqr7YY0BKJwiubJtY/jdKinTc3N2NYODYEldTk6P//1YPW9sm1tnpGdWHAlmr9wuxeZklR/K6R87OBBdV7bUF5Q3MlAJ23ifl39WavsmEoBBQ2BN07liy5hsDYpBW7u4W/t0aJ1zccp38pnebG325c+XaCJUZgTBhxFM4V/+xVaHzeupgM2QA4OsVX60rm/Iihcpg5BOg1F0SIImnMahXgusq24JAf+KbKIbb+PpRz5G8Cj5gcFnUTxZaTRORkQL2M1uJJIDZVbYf+HBqshHf+0+9AJON9FVyct4hz4w5a1Pmb7I9AErCs3m7cGARv1TBDvF4pB8eCLxBYd4LVzq0dPSNABoi/ovaluUxsvG0oPjXur4VnJZ3hIjRtt12+EXNzOjR8PNZ2FIgzjaZ444/ogwgmV7qhQhGz3uZFe4PUby9WMqg3xDT0BZ1O+XZrJETpSYFPjuV08zMOGujfAvbwl66oh4akDEun4RLvsIHImLcVJ2ISZewxZ+jurT3tI5hmBGWuSnD+AD2gHLsg2eLLYjM+lIAlwsa+vRWSFhZuayioRb8JxiyZVYewt/a7OKO9aUN3oLVlokXDbezI5yhefiIZV5iFhg2JRjdM+bZ3iWR9Xgse88lqMj+s/SjYVhXeD/MhViwdtePuzAs0mULVFNdi0j1x6juPczRUxoYPn1RacV9OLKwgilNdkU37HTewJ8zt2jGASVVgSTKtUan2/bqhu6AmKUR+t3tRGF4YbbsS6BkIICV/Yg2xUhASds+0KdWcX/QWBQYoPwpnb4l3wpgNOfwqoY6NhyYHgOPVDGO2sYEUrRERT1C9j2dAEyCoy/Vnr+xwEIfMbKQOIKhiLlmOlcocQmPfUJSodOwO5YYMpzUbsjGuVw8tZ5ALgf+ANsTqeZbyyOejrtOE45bEB+iDjhEqBIOZxMvNybY2Ohs3eXSbe7d5ChoKF/OElsUj7D3RcQZgSB4TEW7gXNv1LEDFJAalThbhgz89REagmHqunt2EYlj1oXHqHFQrOeUC0fZ41fwGCuogNk2bAjT/ZnLA4HG5kN9CmyIsuvcDY19ZRp147sPy6UUl4n21MY8WYZDhDUS7odeBm3djIviPVn+d76R+L+CKxNdHq62jS1fEET1Qtnjf+yb9AqsZqIBZ+p4P/PkCmG0AQfOtUnHfE9uzGnUGiOOuIZ4yY82lcRo3frfVDrbhWcSpld+3IT+7RscjMrqyeCIgSJNDwIaqwSC4WI0nU6+hyPb5pwaGzklrUBYgwOlHb7IH1OVt8+SP2bR5YvcfvcyBYbt3wn24CftwWEAfSMkRJyj+qTXvuiMGyAd3bHcXQXstdZ5vlHQHjz7AOuKO6QEIYAoTl2Uz7k5Rl3KN0TC5SsVjGIgjKNj53cQq/w/omrDPrKsW22cF+i6fJpVcppDrZEiW4LL3rylKgfWKlFo3ztFfoqOol4guxMV/LmSDuWvfH+Iw4cQDBHQlKyHIaY1TOzn6T+lahnlAa4spXx/LY3huMQk/07C1lWqm9rOHVEqlMaDMS7HWQaYiALRH7Qjc5jQV8TfxyN5BQyUGtzDadrOwO8qqZeHkzkkSRRyVyNunIbxkX+QGgIu7lfWzSSiwcTZsZEFvg+4wH8kB5XePcX9g8WHM2N08NfjUkhq6i69k7ZC8IetFkkCn8rF8K3305XDrYy/24EFB2KgA4lavQEsdKMI8UYrcpOXrNYLtRzOcCXbIsSYSOZkowbZAi4dOtUJsggsPqX1yn2lVNoVHPvcsyc9fwLkhVGGLX9+47rRbddeij14GnsJaziBvw6IzwIcc9XSAoqooRN6bRJo60KKwEk/mOOgcjOLJkMitvhWWW9D3lQIM0kGDuWVzTH8yuijbA07OuISlyC8wq9J+C5FdYKrXMfaOHdja4teqP0OCJWb9jbH3TidL8Ppd5QZMKwlmybOTxJdrRc+gX0ZlGXxJXkGk23mplpaaM16hCXuvObNMS2ZTSBxDEikI1pC9dUo3TsHGUqbemRQDrB8gWa2qzCZpxiaFPp0nV/iPvdhj9zxr479SY9nYbnYuso6hvOjMgIe8nvi3uqLtsxTe0gRkOSzJz394lz1V7UUjvPb6b3Mn/xrETWU3DerYwAN1XnXGYTpXRsIyOzWqYKcNoqMJmtMxJbiHi6tej97B9tbIBhoSx1BXVM90sZctCO/xKfLGbOEh1wOUfF59iHEI/9G+aVZ5u3gWq8sWflOIawSpdsTxXHZ1l5Lk6WBJgjNZEP0lBOcdE8XNIF7pY0HeDIbL4QLjop1M1H2d2+VXgL1yIrNqIQ+6wMleYz6ggPZb3Hl17A4MTzh9ca1fPDFyHm43YsZJ7UdQACAzd7EH5meVyUgnrkJimypgOyleYxfh26uAJZn063QEL/n068Wh/iYbEmaHZhl/9d1Yh36JRlIBhumKQ2V254qdPIE6JISagdaAPOuxY97YxZMeVCaXPNZjeI8gAwncZ/5O/kOuNHqhqrV37T5aNo4LpG5XcdyaYxcOcU16JAO4mEl2rCIAh7inXREoEHSK46GcZ31I3B10+OES4aB5RBvn3KhP7900OM/Ordo2A+uu1MVBY5fSs73025yzBoF43uNS7MEUQjfPe4qdAOURtPGb20mZg8XHWOJD6aJlJVNDS9/OY08f71D5wMw8J5hmbVforMAmUZW6Ewd/nijHBKhyVlMBnrJttco86yWEdsgoSr73pPepbs+txcbLj/tT805WIezDwNO3NKiAYQ9LaXHsuw2OhmiPsv4JDZxvSBiX1ipF+RZ8FFQWP2r+m+Yind9gm+euobdjAjaA4Z8ZI4nG7zNRWlyyOd0Q+xyZ/Ripmczv/HF0GtgnH8ixi8OfcWfYZU8ZW0sidjFncw3QqYO44fMC6goD5YVQ/t7YgHOXLDeMQRZfTN6oyCWCXrNp6Jjftp14vZxKQ5huZAtyWOGFHmrTsveR4zSdQ5/RCUuCh71I8iUIaAHjodRpev3lYTeLW9Q/FpT7jxEkKSx+xxaH6zOjdq0MM0MaAfqAgdXTqzqcN2vDoPEqL9u3gnvETUbdg2SAhchIb4Ihan9AE8zR678VhuNtkpRweAbyBLJy7QgyoZLQA4yiE3qsI+FiUJz6mY9ikKnsMpL96ECJZlkfbKkaahoauvq6Xm2yknjqCgFN+YEYeYHCrN+WxNeDArsbEsDRQax/klfP0fHq2HtQo7SPjpmB3K0V9n1XOj74YeUMpOulcrcLBtg1DmzoM5pmfhCzwSZa7AKC6gitu8yaQ8hcUdC6NQ6LGIrircPUOmx7teqRdIIdIwRLLGp0AVjjFS/AK70CUFs91NkaFJfgCENPnpF+W+Dwz/YFY3JVzIkX5yWJFh04lLEeXx9DqjdSNIsEeb1dEtCSHVrvKUqIl/ExTjrUNpSpC7N44xS7Hh9jvOpUvdlKGWwrTC4jKOkNen7vdH+Raf4dlZyYBrX6lAkqwoFRNCbd2aL6csZX90kg3DqExoErSuG5bmyBtuPryHxtqVtiCoi9cUWSzlHBbS9sC50QQfU195NSwlrxQTDcUaoI+Aka4CWWlQQZPO96PPfSYWYfgJfHmIVINxOruw6XPywIgDkzZbMqA2eJyAMdptmVyhfGXPnhLE0rUGaIkTtl6z5A9hH9CToBb2MVgFphecdAu4yh93jM7OPnhzIDIpwvsjN55vSjV8Ryiifu4Tb3x7dIwJOpf1xHrwlmRrRlBBcMsd7oGnhRNNtQcwWuLu/dx7I8vLBOffsyPUIPaTeAXB//3nGB31hxv2ilKJAxN3SGZWytl8iKydCoMRL9JKcWl7OaU8glUUqKhNbdxIBKBkTePcU0O4YM5d1hoQHr+LuJVtkeehfN3bWY7EFSqBqsjMykOeQSykka3fuVOouUkw8OzygVzGtjDj4oOlAe2HuWcjjAK3uerGUBpIXgecbiUHg8T4g904n+umJveU6RqRKsM6mGtKxqQcT5GIJW3A799nKYmNIZBRd5sZ39Tz+EbxaRjI/9A+xHA7zDx/lKu2BwulSM0XT4y1G9+AlzICDrBf8HCY/ZgvTPbl0lbmAcHn6/L8bVjdcYGhzaQM7/atM/P53VcaJ4x8iti/OXV64g/EZXSXUjVMvhUJfHKxppWW3YSHcfgUzSLbgKxwWgr8cDvxfRxpQnrI3lOsgW+YZk1HVvEV0g8PBCnHO4UxX1PwL7kZWRuolJN9xGEXR1Lp2Ce/f7tEENFWntFl/ckm5fRjoDSz4KEKxlDCWVFTp/IrWZLmlWOt8jdN4+T7VertMsoMVtGC0WaEqkr2qGRFdAilCcJ/7AOp2p+9tKSiOs928WTPrNfoFxewecDGsSCIohdN7L89gDje1QZ45HnGZhN6x+dCzfUDor49vfH1EnDlhu7D1oRfDesBvM2/Jl0zWiFUHNDkhROwobRK1i+MmXC0q69QN/2mtDKk1WctUtWslamSR67oSLTI0D/nEu4AYwcSesXLGoT0UQ/miKqo8h4mLQPcwtuDWWmmeTHZCsw13Wv3C1OERqqtyoSTCCXjxLKKe3+L1hwk9HTd48/s6z4RwcvT6EWriPoIzD0EPAQlR1pFpRFegKkp/LkYPqfeA/HaDEe3xeU5AWtTYXBVQCHYhEzK+g+qPnLjwLaF9DLwNITYjrCxDayJbCYN7y4CUISLFhKqh1fBbCnFeSREqs5pwVdCh1Hf7nS4skVlvxOcFnYm70EXKvlOSSJVV1LLRlGyRd2NnNjZoxshK76O01RKMGJR2wgCGMrNHdqNOKj61AdW/kxOFx6RFBm2D7lvEXtvKAwlQLSS4j0tsJ0cI/0DSrpI8Hiqky/xlXqtxxcm99UKbKjn6R88CRlXQOqM8Fcx2+Itn9d8Uk2qXX1LoGZKJf/1k8gEecO1aI2DQljl9H0ReiabnkmxEhukTzYptsAtUNNwVJH71eKxUQXlcThmYsq8OPH6yEbwGSxdnSHWWwLa2yPmZfIirpTfz5KTsiIJNn+Ee0zzE4Ma6x8eIEbtEiWzBJ2vS7koBNgpw8XZZ2Cnr/0hcURw9XiPSWnW5K9UJun2j1+m+x6Vb4fexR6x9m36cgeSlD1O4JlMeVfgsb4kEfuq58REJKz3FzMiDN28vbIUxYcataLBBPD/Qk3oJiQA0k6Oqvgv/yePlukGPy99sjxWbVkr4R8+FrNMI85+usAo/6HGwT/b+PCSsRfHPa0HII3r10X8IsRKj9eu5dUwfTZLfFOkBIv2sUP9qQ/HLfvnEYlO/7LknsT7aiPrV7u2h2UVoi7dy2cIjQYLCvvFm/VpJZFhc8q5W95Nqi05Ra6cAdI/vwTEE7k4k5Kto/GFE/JZgMczZzvhMk+xgM0l91dcQoiWg7j7OK/14nZ5bmPft8aFLmnTszpT3DWHW80M67R4/vf2o0YMGFMuHGnAPLOOCUH+uvwrcK8UKuMW7brs+nUxCmltkPFfhaz26/1b/lh/i83y3BmJGkL/jT7zyhFYrW7itjY7DqxJSVU+IHzfv9F/oQP7QSOO//3CVPnYpQz1fL2d5FMx4kiVcWzIlrGXZSiWRi5tcC7DSP+OXlY22TIZdXk89OkoR3/3/AvvR8jDqXnz+UrJECMhHAdwvuPOUr6lLF6tAftSrY4RXmTlCcUT4sqtc5fSnUi3izJhzlVIJI4U8rSHdYgyyr9Atgz5a1ntZx+eaEEkyiw/F1ldvCWill5I1IVWBUu5s3wiCiOdp8FdIsrJ+eQcG9cgQEsijxQLnAhzuMFhid7d+ftYLvNTShu5pLv2RaKrnIKwAu8y51lQooIKcG0cy4JehBbKDBcuoSZOuxwTVBS1HYpE118VpusQET7Sm1K7MQ6TmUe1LQ5bOScpXgppGX19KYa/+3QFCVIDhcQ/YD/SXvh4j9GfIByz/07ndDZbRKYG8+0SJjUbacA7tTE2zdrlrxqffUsppzEsLZDzPuTRDIRpQl3HOrjZXh2GNSyw6oH9jERphBn+p2ZWg9BfGyCGB4BebQIleGzi8g525ZwLS+GEqlILPoJCFN5002tAPKtH+HNh7wdm3yE3f2ZxFInjSJCiHdE8VXPkWBeFStsgvA9bDgfVj62WM7D7/LURmrA6bQc8+ysocSR6O9BbYn0IAxhgcUqlryZW5+bIyteKMKZSeZ1HrBJH4/BV9E41aRF9oX5THj0VdmeJy0CTpbkpTVxDqmHw15iVSUFc699FNethNQ4O0eFFmc2JTzF7FuTF2W2ag8JudJqWk05r1TxLxxeJtKA7ErJpaiAids6oEg6SKmptMKkTxohZbTw0i/vBibmD64Wy1Ikg5FY8qfqKoX6ziFg7zoUGLugR3hESOEP7oQqIwJhp4drgW0J+Bv9XqvL45SBJEY8fHIVb9jw+Pc19VD2g0WIGmICxXy8fHalSr+rYgwtMF4SXJ+NUfK9C6wm8oyaUBgCWNfzkRiBM8+qKBZH4D1m+TS7Poq4Ecftoz2XAete31S3yU10S7fBJfbYrr+DbkFQCLiPLt+Jw6m3KdK9oz9q2UlIkmAOsGlX9FV8hJGO7DclAYX1glMFjWIVyyxC1++lTbM/Ye2NvuUs5rc3MRQV5fzYlhuppmZ+ZR9Gc1pniqGozh9wLc70D+xy4v5K8LgwFi7HKwUyZPRPEn1B3spyqQ3qQTnM/DmRMKCBtOZMDNtEcGH1jmFOOcM2PxyxfY58GUT66/RmEFjgEvRizeMKDvj+XQYyiF1iQ2WDhG6nhom+fwaiJ01LiEMLtk67IUm4IQfM1wkzdzfDxFtqBV7+D3QxMULjUQ7czNvoSjEM30wPHVm1UafXWsttN2NSzolKlETlFnaC28xRzNiPstyU7h60GG648Sm6dqBouLlkFr+gWfohseaimLEXlpDZel0B/7b0cIRXJE48m34ajDxT3QL/4RW4elN38Nm0jm7+BZRx7/ZYsQK7kCIXzfpQFGvEel/5hn8rpNBivHR+i0G+XwLlwH+abNGS8SDpXyiP33PnSP/qVBe8hubemmBLzICdMia3Wg1nfQslGzPKEHgwQp8YaYsiL7J7qBF65O44dnEbSqeAe6+p9SKldBmqZ1th50D1XVD7c2a2M60NKEW2WuFhOGIZADdbVWrqyNvBruyMzG5rR3Ypv6khhpU5b4/ZdXwVzN1nmy/gpej8X0PAMkv8NcipqD+jlaUA96ISaueYNsr61OSNHBp7qvcr3GSax9XMg0t/Bxcv3WgztEVUJtSeK8En+4Q+mEBjnPzBkh6yq22Z9G+srkbyLCCOQDxfSJBynsdvGMeAUtUy9QokAIGdqpirIc6T96AWbLOUw+IE8UsbKHLnmdzIcLyywxbRD4jwjfMSWzZ3GHdR5VKL/fEo/qeRsIuRoqfzn+vKi8IUPoT5HVoK1hyGgVNFvMnx+h4vMEXDpkt+MlNTkFmmvf+pHtni61Wg/9DCzdLO7XuSU45sdyePN1GeUwDj1W6vz0lhCo1QpjlKjxn+8JHZaCrZnid3tpiqW3El0kBUYaZD+b3S6HettUPrzx0cqNlkqnMcgSb6IM2TcJoK7563aEIRm2Y+NI2sPQViQIQEs1BRBt44vJJBW5G3iIfpQKit1S4ggXiMKYxuN24NH5jznxpZVIl75GfAD0YJcMhSyalkVmLj3pTmubnzYGQoBGvStNjSaPtLixT7mIllXAmNY51xfRPUs95gK2Rla/0XMfao+ILldi8Qwzx6q9okQ2abIJS+tZp1TW3OIMD4E9bQ88UZ2L0yD+OlOQX/mVFuZsDhHy+3lEnfjk7B2JnlAi5FeljUe9Xu0qzBb9oM4yWW1/Ydg7Cw9AnQGIl1GWdKMFbdpm+C6I8e+G5jVnLJrzum0U6tsci6fiQS3fPEps7qX5T3y0tWRZwADXyCsyfynsehNGPrDiFsKBTWBZI/Hmae47AK59tLF8lAPhvsqNMkzDHlNS3zjDaw75kMr08ZfbbsO07Ke4R+06bkPJ6P6ZZCvRiylRs8SvnaU22l0jzJ9WJ3sWxM5CB/MQv7gDXDpbznBTudHVZCok8qsLdWdElA94IIe5bo0/TObtIGnpDezJ/Cvc9kDS9oBoGhx4zADs/DTXkOIcRAlVvXxkzQR5xQpW1xZtYUlu0tMmV75JwlJZmTGlqK/V+SDHdIr2SyXx0Nq93ryv2O0zSsBIeNV+rTBhH6Vlifd+K1Teakb9koX4vN+HYoEF+BEHCBJy5pcTmKBUIoH/9uTS8O7NtJKBxwDNVlzPHz1ogfbdH5bV446mEqy5Ex1D6IXbV1trhZxCHa4hPwrwKMSINYBjNTEpGcLLhbuyqV398duWNHdIjs2ywtV+ZkKRrh6Uhu82jlOecT99KL3U8lN0ou0zuNz5xMEImKxQEVqwZd5KFvJFIfI5cwWD7FOEllOfN/4A9BvDEVngDnY2fXoYGmDemPxJHrqP8GmTnDKOipVnfJlex29o2lyZso7IkfpNdIr4lo5D/1rhg5rCO2w76OhigIDt7B9sTmHQ++jDO8A6fT5tBfawrG3nzyV4/zBzKUYD/TTjLkOISFpMi5tm4uar9lZNe5Vb5mEFHf0e9JGmUVEOfDtMC3kwgpzYS5COipcIHNDvcRKEo0PwxshvK+7XPlwovYA5op4Xguz7RE0DZNebk4TfQNWEa2BtR4Ny7ZPqXeCHcmtT2b+8aSM5wy1c32fnVbr1rVWsNgRp+7MaTi0oF2SvZT9UZHoPAIwVGQncmjBf7dzVIZfp3F5zUrGjoc+DSYY4bqxQWTUWYgpJo8cB1sS/K3cN67EbOB1tUBAxg51+T42SuXsb/IfiLpjtNpMrFr5Hs+xCl9vI7D56AhkAxDvtGpiJICaXYvW+iVOrGA3wN2XyntRqkWZGNYb3juzgN16GBFgl6R+Kp3SbhX0k6lPzG8lgmun/OZnItfWhlu2SerLGiHgHGcWD+DnkwwgpO/UmSlcX0wWe4m8DxBrO6IbNgHgm27xBhaNYwUwPmp4gAj94P0T3TkgplkWEnddwq8IItZOHsJUihYvw1cv2JRZf4CBKIkXNmvBxjew1ESYsJ/AJUTSbjdcsOpdK4mjBBwmmbGkuMIMdVxT6dQKWoFpsiAdV/a1bX69ataKYfTMmVtqJN2n3k317Nvyq6kqkpC3VjT82//zXizGoimZAWmU7I/YkVPWdFhciO8SVikDwosdii0tHNyzySvCnjL/Hy4z9B82IsSAgsgpLjjbodcGlN20fZqTaZizTZfJt4jSe/18PbV334acQ/bQekINnKUjh1wA6fMYNypBFsbjueuAJKAoaO108kVhKzHg0wI8r+BBefgzijs22jLCLIxBC5vzgin9RySliatZRS7+3MrmhbVmanYcq3se8Jb+iOlDJWE+DHo41LbmGMTbYd8RlFQPINgt6sPXfK8r4tm7lDHj3dMjFA9doqWeDSyFbFIFOiEQexwRXrHxez8Xo8ylagsiTXX1J2kVniow26AMaeKnR2wR5Q62Zh+7SkV81BgtYbKFxcHBSas4tgddjsoanwbzwpUoIjvhGPeBMeCujjIftBnOxjNntB0/SL88GsbsSd75t8CCiIlwbPFp2ngf/pL47vurxuZ4qVTRkBm7CubeImMi/U2qiLC4VWOe3I2lD2QDBDBrsXAOWVZJP1yS771/mijvy8Ls0o4gzi0f1VRGKOZd5okgh+D1kR2XcBwbNS57iH+cy/0LP5tpHMVJmlwACzfH5GFShR+QVognGrD1IVm0IDSHnoVHwCNG/i+Rct6ls/rUkGCvQ6nKFaJGy2jR3TActF6/X7O1Kaz0WOWS1fhApd94MpDrmck9/HnnRQgaD65FQam4g1iD2Ws54YtTEihgFtYMoesqQ82C05WrJ2oGKeQmKDLWyirZhN/XA3mNMsQ/nCfHJ8haDn9IRZlz0xUxeyJeC9QA+Ura+SqEo+L53NRRoxQ3hLTMSPf/ShcY+hwu8crwjKfhwzAVU9jhp5crdc/2An4FrSMe5tpbQmCdDGAGFkw785B4ulBSjTgDOkZ28x7gURRl4LaztdWioApzNUM4bixxMziW8PgNWexnGl0Tk9YiVNr769taA7o3RwPegy+Fcacmw3ZK57Idhx3z6b8IlK5mD78ULArLmwWJGwy25vjGA1BaTwAincV3b8FRFfGgQhjmEjZpG805pX7778XpfBHQif1wvaqKUaqZrsbeO9WBZWS1NsrHPo47v/VPln34N+tCn2Kc18vT+qDF9tbNO9lDz+4NiljagB2KDylmsy3WZgGXw3CaImZWMUCSJ/ceKKaA5IHuPBOpPxyZuGuZPu1smWOw3CX6qW66Tjta6mC68vozlup2yOxsdKk4TiYnxoIb6fwo85OHbZFtFGxUk4WTNwwqVfSVlRa8jxrp8kdXJkQbcTdDRKKD2WlSNbMydFj6cifn/8AaXH5f5Ek8LHzCCw5hbZNz/cWsnhuHGqgnoNKNZm98vJEPmKbbyGSQx4XjzYd7sm7XQF3mMkGRlvcob3Oh3ijNcr2hXMJPcNaH9diwhD8DJgmj1iQ11w7TYYFide4F48nEzL1XKPa0RnigB1CVUUkIaI7pl2W8SWjh+AkRR9VuaPGP7PXlnAqBYCZXiRbXTO4ZkMPyl8Tf+1TO5YXncJa3hRoNjhqt0jZAm9oEb2XIrof09CRQW3e9i+9zx/9I95u7Vy+7AS1wX0wa9WQ4OsHkyKu4JKjcJLA4xO2XeHPfGhaig0T5oVtGaLLwgW+o5CocL1VXeyN3Ydqi6EqrbjioAEfG25QcqT/hzRF5HYC/lJvhGzMnkV3V1dJC3/xBmAqOsI5xFu/EqJV879pB0Bfn1DDAGMQC00ZrXbCzYfNqRxmrDPRO5YiIWDD1CCSzLTslfqeM3Sb4yOZbZBrBWX0B7uPT3L25MJm03DHl8Pw9VytVl/oHrRXlZAfTDPLki26/9vPNB4GH/WWdXDIuuFzd/HUwl7bct+hh+o5nC1QXziYrxDPmhkF//Lz8v7CU2h5W9724B99lLtCozos0yf2/Jsfny7d9JPoqeO/kx/u9CxxIMBam7/22HFVJeX0fj6051bkHTTv9Xf+e81jdY2jxuPvbvf3/ZDd4cLfAG+PPjfrtgeD/GQ4Vo1f0zT2Uz+vW3FqzLLFvI0cTsb1PpbL5Q85AyJE8Drn+vdt/rJFXzAoXmbHAMdHw7816xp5RpxzxVh0fNOEDPaU1dKPT0v+MP/4FhgGXBTgLUroE1ci5jyQ4m7UWx9dVUQnYqp4KFgNVGkOC0y4qGa3bqlcPeqfu5mULiwO0NKfNQvVXODXnzGnAtV0EMNa/O7GilmK16to3ihJ3dLFfJA837Egua57XBj0Hn0HeWFI8UHZ0h9FbmBP5sw23dON+oyv89F+Goar99QCbFlHC5JNoH3s2gvtPiD8mkZlBB9HjMmYFhcmRjxRL5Q63oayNNZmCBeoP+cniaNDdLt97ntfdm984rdQIg1oUfgx+ne73YygHbuAeBG2BMalBMQ/TOQGxlW/uFjjYvBiNoY1ESFGT855LUivtJAfkqI84O381H/FtgsMnLqg0Czgq3azS9KU15o4B1TAT3Rh+CjCcre3SFImwK/2MRZIe6cHvTjtnMJbVqedw32MMqe6O91/FAi8bmesftMhs7AYz9nhCTK2b0J2dq+9g6QXOwjJCWB+BUrunkwiRPq/Xfp+72tLvGFnHPHDrnnPx8di3eG9fJjw0HSySq5VUI/+bjPp/ZfT9NjxaSW6LPWkt2f+L89V779NJ/X/1p4ONirY09DP/Lf++tL+f8P/AO92fG0OTw+cRQCyO5kDg9Li1xwV7IcAfom/ZqACZSkus/+EI77r4G+JA87/8bv9HHz04l9QWEBOn+Q+fkFx/AbOd0QecPgOcKq6CgjAdLzzF76YBF6+xhPlmx0NkHnqbq4SULTkYHQsgE+50B2MgZ4gKT0dp6ZPBGUlTA8NoEG4Gd2Mfj5I25j4lKihVQurVUKrFVGrhlUr/hccLcFkoDPEDKOcnwlz4EdNqBwXeH6PRnvyU/x856ncTnPiUiUMt1QX+3dRNtgEQj/c4Xu1bR58B0iRmU028AED/R9QSwMEFAACAAgAOUnOUFsx5RUpPQAA2KcAABIAAABmaWxlMjJhZGVjYThkNjcuc2/sW314U1Wav7n5aJqUpECBFlACptiqlJQvC1rpbVN6i6WWfiAikJY0lWibME1aPnS0mrb2WqrVUWCVdXkcxlHGZ0UXWWBdDVP5cEZ2EHVkdJTKgCYgwu4Alg/Jvu+554aT2Lgzf+3u83D75L7n9zvvec/7nq/7dfpIcdkcXqXilEPN3cEh6h8q4wLKH34xqgJcHpcMZyt3A9HVcImPVaZYyaXKAstpGRwvp6fESrYcqc9C+Tg5Vhcr2XJI2SbJ2JYfK4NqWZ4ZHluOp+XSc2ScfkesDKpipZ4W19BfHi/jeGnlYqXShhXH/fWY3jRBxvHyIS5WKuXmQzkd97cfSnNX0voStctBPlaqmHrTyHjhuJLyGu5fAlNOnW22286WvZ524cuwR911tISj+eOY9l8+fK2eyzHMQ24i/LLhV/jaD8f0vo/3naj9p4df+cP96y9u/6bqzTpbpOmNc9/s2nz/+lXNM1/dd8I18dNp9a2J4lnBxMQeGarB+cYE+sfgN2wQfmUC/VRmCLLHQIJ6P1QNbn9Pgnpn8oPbOZ7AvjWBfVMC3pCAH56AT0ngz9MJ2qc0gZ+qBO32VoJ2WJvAzu8T1JvND26nM4H+LxLE+2iCem9J4P9HCewXJvBnDvwmDMK/SNohhatNo/OfTrxPVTJvGy3jyVTfRfm2dBm3U3465S3psfobKJ+aEas/hqf1UvsfUL6G+rNpWOx6N5/aqYizv43ym6j9tZR/m9rhRsauY+cobxsj4wtKQzgc9zV5PQ6fv67Z73BwjtLqeY56V7PrPrfP72qunlfU6PW4quuWNbrkvMFzHM5VdY4Gt6eu0b3Gxa1o9jpdPt8qx7I6n2vGNIfL4/TWu7hK4W6urLi8pFrkKhscdY2NXucCl9PvbUYIZfwAMNniUUC8oXoXMdTkanKuWH01d2Wz2+/iSsuri0uKK9FClbOusa651ON33edqBu9czc0erwPqq/O7vR6u2eFf3uxd6fCthkiaMBd8qG9Zgb8pnLPR6wNnHeXuxgV1jS0YHbSO8wGHc/kDjoY6d+PVin0uP+TVe1v8XNldJaVFQtmP8sD4YPoOvxfaC5quqrqytLzEUVxWDVUWiUIl513h8gxmJlokmlfv9mHzO9ye5S5ogjqP08U1OD3+RjDl9rj9Dmej2+WBNo32WiVU7fa4fEC1+Fz21Z66JrezanXTMm8jcg3eZqdLgUoryc3jcLT6oF/cHn8DNgS2spzh8zfLiUb3MmeOz5szgyspKy0sckzJmZozLZrOvZq8mpqSMz06I/mo5OmdEh9NX70Sa+M0VYP+4fqn4e5k1lJ/hjsZS9dQ7tmn1+nQZh3FaW73EKyzkeKW0ajPc6187PVauf/opfM0L46voHxBHH+WLmYL43jjeHp9jeMnUL4rjp9N+U1xfDnld8Txyyh/MI5vpXwojn9G4QtieQWL+VevL5nMelrL8OPY+1OGv5Hhexk+i+FfYXh23Q4y/ESGP8zw4xn+DMPfwOrPlqWOGUd49DM8e58XYvjhrH2GH8PelzD89QyfTtsvibnHI7fVDK9m+CyGZ+//bQyvZfg8hk9i+AKG1zO8yPDJDF/B8AaGX8jwRrbfGT6F4Zcz/BD2PpLhTew4Yfhshm9jeDPDdzE8ez/Qy/BDGX4Dw7P3CZsYPo0dhww/guG3MvxIht/B8KPYccvw6Qy/n+EzGP4gw49mxy3Dj2XHLcNfx107/q8dYuBbvdit/QLu28SOoJ+PHBQD7+n7ovmR6UchK5J5DM7mcQWQQrwcs8L9ETgyP0OMS0b4IMGHEOPSFQ4S/DvEuHSEtxL8W8S4ZIQ3EbwLMS4V4V6C30SMy1O4jeAtiHH4hFcQ/BJiXCrCtQQ/jxiXiHAFwc8gxqUhXEDwE4hxSQjbCH4MMS4FYQvBaxDjEhBOJbgZMU79MEfw/YhxaofPXEG8DHEqiZ/gRYiHkvgJrkQ8jMRP8FzEw0n8BBciTiPxEzwL8QgSP8FTEI8k8RN8E+JRJH6CJyBOJ/ETPBpxBolfzs/huIbeaP/25Cdj3/ZoeRQzB/wjoLu/niR3d3Kk3zyuDfX6qAT9P+WgmH4IRfYVcfcVtSidEXeHZouqveKHV/xpYOAtakAf6W8wj7NfLd+W/yq+KWmZXCMG8tfhSwVROu5PEbvz1wAfWgIehpbDaa/2UcCqJX1x9YcfhMyGHPM4+YGkV7hbWCDUCNU1VaJ0CQdr6IkfIhFROlIqvRfyQnLnb7Zs2fIutkdPBS/tkQ5+/9WHJ3qmPUcujg8T3X2hdVeIFKW+0J1AdZwyt4+HISuYt6dZIdSV0P3Sh8EwH/hqtnl7iqlLH/hBZW4/DCbm9Sy2qoWdOJ7fxUG+E0+CNCBIB0qgtEaQDgq7+zV26ZAQ6LcFT2jM2w8JyQfnOYOFkK0r7ErLKuxJI97Ye9yqwNHb7T3z+VLVHkH6tFT1W/vMA+bOHZgp/a7YvF2TViztBUOzheB3OvDuj/bki6LzEFgaKXTVcMQkmFELXcN67uZnftK6FJSGB/apwahGCOxWlwb61OhSMMSjmTlgUjenS5M184fWJrv0YbgQWpf0106cke9iWMIuDGn3UU3gqA0rApdT92ry8vdq9Pkk5NDAZWg9WBVE6fC9S4TFwhJhqeDok/udeKcB74o5YS+ff/IA7cee1drAbt3JV5jeCx28FO25tyBJOk2Q9u4O8T3TCuzSuVLpXOiDS3KPHbsc7bF/ACpwXG3usIE3xR3nze35ct+lWNsu5JvbJ6twvI6YAnPcvH2aoQCmjbldg7o95ZHc90/eiuuWEBiAHv0PjPRrvqDr1tKZ7/uLwIY/RQUF8v23QtEpYmC3Si6+h7i2XzoNragWVfugOO/3wVnd8lZgj6pAjwZa3m3o7SlWCTP3mDufhQLSKRKSKH0X0tIwrrscDfnQxUikTPoGI9p3EdkTovRlmfQXVNuOzYGdEVpLCkbbuU+QzkB9u7/W0S412/VJ0umCrml54ZnQl71KvgbyNejs1fwMzJd7cEac1R1kvS4CBegbUTpXFaomGqft3ZZQBvHuNPg5DlPOj+zo9d5QHeH7xMBAxJ8qHUFj4aZLdDyF1JCbG+nWBy4Y/cfKeoY4YQkuzg3mBnc9R5aV6S4DWV3+At0UagXlk2+TfqmugepF5xHpCMxuS73YrcnMIo1ot6bnBmHdmoHlJA3MU+0Wg5yhh/SbNJ3acahlbmgEGBSk99C1n0tfSqdDZy/AoBmAnvxXNNYNxurFqbJpcwe+FIQBbYHASBS/vohRkHEbOgkFiYU3FQtvEAtlxEKK7Fy3dmMWCrB1gJNtoaG+ht7QpQGcLHBRdX4E4Uf8H+u79blBaBB/MmmGpUTkvwfu79qCFd4P9QgLpS+Fe6DSfFppRykulK+jrZ7pNyeTpty1Hd+3QNkuKAsGbwA64v8zsR+yQsGljvAwZX6HfNQPUop48mfFkxN64snHROTPAmuh46C9cFF4BNb6KqQblAUYRscPovR9FVl6IwNkZHSnhzKhOjK0um24Mkif3bsknAuLbG+M/h+i+ucG6ABn9FVEP/e8rDv8gjxj/g00d20G0+8IPPYwLAF7IL5jQEt92L/XEQNf5h4i9YOdnT8w/oYe/J7GfVCOu4YG3ScHfSSJBJ0JbRc68z0GffL1BuZ6E/UnMCD7c/Pg/lQP4g8v+5Md48/J89SfL2L9uVf25y7ZnxZAodpB/MG1M/fz0K/QirTnpL67mAsMaFoW5QbfwfClI6Fd58mQ8R9/RdUyNbBfBdPg5NBo/d01XOCipiUrN2iXjr7Dy0Uep0WOdgRbtIE+1ck3QE1ez+H6eRiu6ycCl8wPH4dxvi0TYxIWw1jvw/laBSHAWp77Pqy4+OCaez7kJr4dgdQ4OUU8uzf8xnlcn6pLpcs10idwAfir+Oi32CllMGsLyqQyq9gehHU3226tMGeXWWvN2dXWFebsjVaciObszdZeIl+zbiJym3UrkW9bgyB/vFC0W8/g+2Zpf6n0FVkkduKDl9hjt1rIFe4dGebfBI/n+zRWfN+xz27V2KiN1FAd+DtP6pfCovQZ9Aw/kagXQgdBf/vP4cSC9WpBqXQFLmnSJxjZ2RrpQFWZ1DdP2g/B9eN71KB/FoZUiyFBNM/SaDbSaDbTaF6j0Wwj0WBUB+Ojgj4TOz73N5X1lFlTSTBsHO3WVHLdy1+USQJJVQJJx5BtwKQpjCXUeU4ehMVSGCIUFojSn+Dy0zP9+kxiwqsjMfaehTF9GrtHH9oMJcq6camV+sJrYFlQnhVing6uHdeOa8f/5yPytx93wG/2tOkzbs2bOeu22/Mpy6l4tUarS9InG4wpQ0zm1KHDhqeNGDkqPUPOHz1m7HXXj7OMn3CDNXPijVnZN918y6ScybbcKVNJtlBYZC+eUyKWzr2zbF75XRXzK6uqaxbcvfCeRXXLnPWuhvuWu+9/oLHJ413xs2afv6V15arVa7Cw7MfNk4vqPB6v30K+sFj8XktDPSd/VMhxcj5/vdvDUY1mV7O3BXQyfYk+/yT4mPOTXzPiPvEM9uHl7/3g8lNfVjglGvwGY/G4VlrkwhbMtdRm+mqVfh1EDwz9WO+na+N+0tPB8n9kV26YTJ+lINM3K7PekpXpw/enQGTJn7cs5OuMJbP+Fuib7Dg1TjVGfdtD9F1s26lIZAO+Yz4diQTJS+5IBJ/Rg/8ZieB379q/RiKrQPaDxDu2NriG4ffqrSCngewHuRjkCrjQP4u3QnC/9QHIXpADIIMg0+G5uBfu7W6nL7uVd66qNZWcalWqakxKkr5XJe/bwPec/d9GIlvxsmxKnWNKn2s2rtS3cbNH33bTVOsEpbwd3zd/F4nEvN+G32J8v46xYPlCU+pTfKEp/Ul1ocnSoyk0Za3VCiZbp04w5QWSSkwF6gdVBlOe3WQTTFmgAqpQpNCkLzFW4Ptg/GYE7bGJ2noSbfWgrbUaKNCJtgI6tQ1sZAlMacHIl/+Iot88LmN80LbkGy3kreUFU3qn2m6yBDR+gyldIOpzjdjpWVDvtP+KRNarYnSLULfUVMB7DCaLXSlSZlwNUFBgkbHUJM41VbhMot1UAOdKU8VckyiYCqpMYpmpwkHScptj2zdC/5LvzFC8ky8xpQfQp0Wm9CJTqmDSEz18wvX/D3rTMD7Qewj02qN6RVTPCxHOJYrl4F9WqcnmIO3EkXLpMF4+/jvLYZtWQLlb4D5rbWybCthO/KNKi8hj6yEcm2evxhDg3SY9s91pI+RvgvwPorbs0TafazqjysO5ATplMO4vcMq4KMZxUYzjwo7jogjHxTxTL8+/aDBt4EXTJn6R8dp18dpx7bh2XDuuHdcO9lD21Sj7aJR9M8qh7J0wjh+cnzB+cLuz43hl70d5An5ZAr41Qb3PJOATHUr+ndRwdO8E3cyg7Jm4hX7sV/YIKPuQlD0nyv2KsldA2Y80Ni7/3JWIl+zXoPUpe0c20c0uyp6XrTRf2fvxGMXKLYuyhyG6V6Ugtn1W8LH39cpeFGVPhD4llt+fHOtnAS2fHFfflYjs/6rk6HO1l23HMxRPoPkXKP7fvtVS9rHHHzNoP8+hcgGVDVS2UtlJ5ToqX6ZyO5V7qfwjlV9TeZ5KHd0EM5LKiVTOoHIOlQuobKCylcpOKtdR+TKV26ncS6WyP6ykqGiWJcvuWuau81jycqbm2CbNyJYTZE8vz/S2ste1Vj14G8n6nxuio4NuuHFofkq/Z6yir+yxfT/BZPzUoOyt0qtqIZFyCk4z1IM7xmuS7gDn1ePBtvrwKAyX1yfh9nGNXotb8XmVXg3ldeWY5pNw4yCv0mk1iLRDQSRptG6PH0ppLysTSLMuA3fcq3NgOmrWj8G0Xv8CMaZLQWOYTt4Guoe1ufocMGmYCMFopyaRwW3AVUE7Ixm/AOoNQzDnVgJSDXlgUZtHQLphPMxu7UwCLAZccbSzCMgyrATT2tsIsBl+idZuJyDPMAut5RNQYHgByxQQIBoOoGmBgApDFYJCAhYaTqE1+8gS0K414C497Zy0eTxuF9GiNZF4vcLwHVorJcBveAxz5hp/jnPbMAfBvGGL8GWEYT40gLbcXIerkKEJrd2lr4KiAcOnaGC+2Qmiy9CJZe4ZIuHFw1CD4N5RHsjZYLgeDSzOWI2PcYYv0NElxJ1XDPh0qV2ajN221dCGOQ4DPnXuMJzHnFriW9DwHOYs0z8C3HucegdEp0vK+DWuKq9ipDebUpJ47LFCcsaMIdhh6JdJng96/SwAKZUqTLoxOUaDST9PdZ9SdIcD+xJOWGTMw+CUJJleJoax381OWDGTniCMWcB0t5y+De2NTkGHhzwHxOdkA6sO/5/E3IrVjhmNlHkW6l1P9MwtyI8jPIzG73G8pWpuUuMg13yMQq+ZDGLodog/WauWC03F0TdpWAZkmH8PJ0OOnG5CfrKcfhXTNjk9AwdobiqWVX8GhHFHRosaw3kQzikFMCSHPgl+GOvUqW6sbDI4aHSqhxIwFAu4KFiCLd5A1UaDT8b7KPgKDbgp2IYGGik4iWpNFNRAZUYPBevRtJeCe8Bh488owC9GxpVqHQFPIVhDcyZhPQ9SYMd6HqIgEw38nAJcdowPU2DDSh+h4C7MaaNgDHrwKAW/QkcfU4LDegIUvIM57RSIaK2DgjdQrZMC3EphfJwCD5bpouAQqkkU9KPXT1BwHj3opuBdzFlLweNooIeCyxjck0oZVHuKAryqGXspqMUyT1OgQkefoQBPxl9Q8AVae5YCI5Z5TgkBHV1HwQkE6ynAyWDcQEEYK32eDgr8hyHjCxR0YD0bqdocNP2PNOcI5rxIAb5MMr5E1f4dc35JwZtYz2YZDMFpeKOOzsvZnG4UpIfhViuVeTjyyeo2elc6BP9to1JRfYTTFaLqFRW+Vx1bSVSV/2IZvhTZABlc8f90otr4zzB00p4nGhNQQ7kdG+H2cKr2dMzGRbESL4K/QbWO1EnkWyVciZG7CDGrOgmHb2pxHP13b98CHkdxpds9PT3qmZG6Rw9bD1uyLL8BjZ5gg2Ujv7HxQ9bDlgF7GEsjeUCaETMj2eZhjAkvGwc7YCBgMI8kZJddAsu9gUvyXZyEEEjIGwhhyWU3uewHfNkPCMll+bIJ9/zVVd09PTOSyd7v6nO7zzl96tSpU6eqT5+uqZ7uA+0WDeL+jvBTdExDwly+tey/kbw6j5llmxaX5MOs5Dhhw6BcIclHGAXZ2hAd0wYk+fZS3NkOETZBR+XGaXaiFX8LSeT026i0/M0AqgyQu3jpqNpId9n9ziW08jPqczSBVY/ThT7i6KGjOk7IFwg4CGQnId8m4Dk6ag7zsl8k4gx5qwc3yZmXcuKXifirIlzoky+aIffiau0Av/g1ujgzwZGnCflPk9OTJvp+B71CM+nXUnV3cfrpgCm1lXjHKmzeTs2WcQ2n/5jo/Q76XRW2duOaQzu6y9WerHBo95hD8n22hJr3XVq0YOLsYzcU9EL5L2DHTcyZ3L8Ykk/C4qYzbWLO9A6Z8W0vd6YtzJkWk2+0qMKZulhn30w47oWmM21ltNcI/KkqnKmbOdMGck74uelMPcyZ/oHwr4GWluQ+dvdUyNZ/8pnutY3JeoYop4pM99rOKKepJErDvfqZe/VhKQ4d00YkeQejTKPGBGGqb5bZpvq+3zb2d8rsTviNg/7bMrsTygN2J8yY8BSzF2fFWNlYp9XPa5nXOtckSbXvlzl650+OKpcE7N5ZWJHTO3W44XpmyQgplENEn/WgygjriTDreYnBWFg0a5GXwe1meHCUSCr462dDwosIEaSGMkwjmEirfYEGzBI4gt0+3xz8hgpH8ITH3/AKicIR7PH5Gn5NAI7gTZ4i6YlShLafyFZoe1eRGdq6Y+jHvVbkKoJWM9DlwW12CKtkhbBMhHJ39S7ACGGVe2YAVjVvHaLk5xnm047bAW0RYC9iaO8blSthgZmkmNqoHwF8FHFqmMHeeuLxvVV+PQJgjE/fLI/xJhOHy/7pIM02Sdr/QDw015RxBeDW4u2A/53kFS024ccBL2VwoMi7lE5lc/QlMCvL+Qf/xcd+NPQ4Fdfml62WUdMG+r8YkSnCyhIi4O1YERC9SlvmP06A0YQCy0tBlEJD4wlJu7AUYjXjHZKodZZuZw/13tPgW6FcLsMI9J/3WVxeW/0EwaXPoL9ZmO9v0S5A9IQw399WhMKSgTDff56ftc9AmO9fzJCQgTDfv4QhVQbCfP/5DKk3EOb7L2DIQgNhvn8pQ5oNhPn+DoYsMRDm+5cxpNNAmO/vZMhFBsJ8/wqGdBkI8/0rGdJvIMz3r57egzuZgTDfv7Zih4JV3Ajz/RcxrccMhPn+9QzJGAjz/RuCeKm5z0CY799UhiFw0ECY799sYJwcMhDm+7doEZJ2o4Ew37/VSBJyq4Ew37+j5CqE+Qas7b+08mq6cq+BMN9/WTVigIcMhPn+nUydrxu4vfl3+dGDTxqY2vyRwAMI8w2E+f7LmW6nDYT5/t3aEQVhvhdhfqCo+mmqohxhvv9svRRhvk/bjBuzhgsVcGPoNW2l+ZStrUUU0c28ZpzFeHh60w4qnPd+wUthPmL74G36Y0waOltDbB88zCgaYvvgERNGbF9cw3y24lKS/ZaC8Vv0c1xDbF88owYkDbF9cZ3p54jti2cx+mSxvR+xfQnF9qwQYnu9sawGzwiI7fWwCSO215tMGLG93mzCiO31lhDKehHbG89WT7AGX0v/lyK29yO2Nyi2vwK9g9jeoNieIYjtjRhHENsbQ5wNsb0xzBHE9kacI4jtjRGOILY3RjmC2N5IcASxPfmLiSC2N67iCFsNRrE9QxDbG1fzK4jtjWs4gtjeuJYjiO2N6ziCScI4wBHE9sb1HEFsbxzkCGJ74waOILY3DonGoZ4bOYLY3vgCRxDbGzdxBLG9cTNHENsbt3AEsb1xK0cQ2xu3cQSxvXGYI4jtjSMcQWxv3M4RxPbGUY4gtje+KMqA7Q6OILY3jnEEsb1xnCOI7Y0vcQShgHEnRxDbG3dxBLG9cUI0AYrezRHE9sY9HGEPuvdyBLG9cR93CsT2xv0cwbg2TnI2xPbGA/wKYnvjQY4gtjce4WyI7Y1HOYLY3viKiVQ8QBosLOKDUZOKqgiu/CPm/nfqQPcr4pebVQ8QKn/kc/8SXD75r4heB0coRPmoFjDu4TfRUbOOs34/RCG3FWtSzCE3U+Rqi54hd4PSLzdLNVt5mZ9TmRezy0gzI/zib0LsgvKx3IcrLRhqj1BLauPQ/DXm1e5fmssnV1I0VDfIOBaA40PS8fcIQXaD9jqbRVuCkrSIDizskd9kE+I9hB6lY/YR8vluP3uI4T/FbLiPFDmXVyRjkYv8ZvE/EtucZ0JmbtdHx5zHCWknIEzH3A7O/p7ZBorQqQlKYIWHTKIEsI6kppaz/IlYfqRlNbVZmjePX/VQTDN/O0d+Qgr+kYfWrSQJjDF+zSDGucc4MqOUMdltIKl+5Zf1Atnq0X7DumKGvMOvvGGl8Stwr0frmafMkIpgzdqr4BMvMXu7dwKQT2LJvGnvl5i9b6DmX1cs7P0ys/d3CX+umNv7x8zeC0uoWEmWveud9v4qr0jGelP5x8W1AW7vL1Ohe0q4vb9PwPN0zP2HqintfbTKtvfSgNveJ6oc9v5BlW3vQRHucnu/VuWwt+S2d31Be//MtveTlr1r22HaF5hp3ZsdyCc7qfpqmOyFWWwY63T/oaPuZg2FFoI2SPjldNREeekFpElFULSuV6q5kl9opwvnBIV1pJnXO9rIH/lqt0Gb00yb+hxtfiK0Oc20eZRqfcDS5jTT5mPCP4A2LQ5tHnFqs9ShzXMObba4tKHBvgCPuPKDJczteAB/j/W450HVTzOOBvhTq0HNM7iGDzaCdDuhBwyh4YPngtZEnb+AjlmDoJ1irvgw4cfpqE/F6NHuIUbbTvptoaMyDr6HDVTnOd7JvLVBZzMm7+maH3HdVhD70WL7eektTt9M9MeLHW6Jqwv9/N3QpXR13jSO4Bd/NXWVdrnfucvN66u03a/m5krb/dpLuAt2FnTBv7dd8LfiDWPNnY7qukvc1T3srO4VR3WPTl3dN/NUN+83ToHBKlvgH/9GgZVcxjjeyVRgtinT+fzlMeezBoGHpCJcq/05ZqajzM3dO4bIJ9PCiY7WgOEZEvs0HXVzvSg0B7T/IPxDOuoHwPdF5jCX0lPz9jLhRMeC38FDOOHHy/gcd6MBr+kXXlM83R4IL+jWbCTVVPILB+jCW7o1Qmpmc/oR2MpBb+L0e0vxNCRSOmQvLS1p9XO1+vnSwlXTHc62abpjArt6umsC6y9o/RO29U/XZ90w0DRmYF0qgvkW3YVBc5AZ2L1Vi3zy0hC/EdzADPcpGQgphwbQVX7rl/G7BvkW/yaizdlL/7WW00McHXNGCOknYBsdZ+HXDWffSJQbCLuGjrPvIORVAn5IR82LvO6vYuoLOaOMFmnup46L9gXcVT3MGeY97rixSuck0KQEe3CQcX9qRAsSC4E3ptHliWbA4QVILSVZy5pSkjxmPsw2I1HoWSQDURAotfysCIRyFo1cxOBZ29kb8NatxIpD26K2PkVnHFpvoLWRzji0vkArxMkQOaO4Tex+YAz5itqOIjID8pmst54mP8ehzitp7SBuHKFFvrabiIbDuNOjtX1K3DiMuzxaO98ORvKVe4rb/olYcPimExLEExrY1lEtXQTg8EXl4nPfJQBHm8jFGDd5itpFysY44fG3vUlScPgScqj1IGmBo/SuYuljPxIq81UrY/O2ZmZs3G8xU9iXBFkTpVE/H0bUsP7VC1Dlbw1B8CCZooQZC3/zqPG8S5LlaFh6R1uGXVEWzjiE5+6P6DKe8AN4MYUjmDTXJnyFmKp8EtYCl7yCxAsDpZLb/QL2lmgMfhtRNfspngmHJNVMJ+maamaZVDPLpJp5JdXMK6lmQom9ti26u/ovgJFQKrpnBmCDpUa0Fu1RvNxCakRrK7qMrZaAZ2rn+dFKLYTUiLaYIaEQUiPaEoZUhZAa0c5nSH0IqRHtAoYsDO1lD9oMaQ4hNaJ1MGRJCKkRsg+QzhBSI1onQy4KITWirWBIVwipEW0lQ/pDSI1oq6d/G29AQ0iNaGsrvos3oCGkRrSLmNZjIaRGtPUMyYSQGtE2BPGL2n0hpEa0TWXvIjUSQmpE22x8gNRICKkRbYv2Et6AhpAa0bYab+ANaAipEW1HCX6vdCyEYaRdWvk7vAENITWiXVaN6h4KITWi7WTqfD2E1Ii2y48+ejKE1IgWCQwgNRJCakS7nOl2OoTUiLZb+yN7A6ohNeIvqp6JlBVSI9rZehlSI4r2DPsfFwJwPujFnUfRvoEtr5AaUbT/BRCpEUV7z8N5vQrnLSfqxUimgOL1MkuSIwXaq5Dy9yBtErhNP0dBaTiCB2mTwGFG8SBtEjhiwkibBGuK0YJAgAhRBcJ923ANaZPgjBqQPEibBOsYnwdpk+AsRlfNtIlqpk1UM22immmTCqRNilXFLIS0SUlj2SkIQ9qkJGzCSJuUNJkw0iYlzSaMtElJSwhlNaRN9Ger/8yaI5MmZUibVCBtokeV0AeoDGkTfUApZQjSJnqMI0ib6EOcDWkTfZgjSJvocY4gbaKPcARpE32UI0ib6AmOIG2iJzmCtIl+FUeQNtH3Kj6GIG2iX82vIG2iX8MRpE30azmCtIl+HUcwpPUDHEHaRL+eI0ib6Ac5grSJfgNHkDbRD4nGoZ4bOYK0if4FjiBtot/EEaRN9Js5grSJfgtHkDbRb+UI0ib6bRxB2kQ/zBGkTfQjHEHaRL+dI0ib6Ec5grSJ/kVRBmx3cARpE/0YR5A20Y9zBGkT/UscQdpEv5MjSJvod3EEaRP9hGgCFL2bI0ib6PdwhK0YuJcjSJvo93GnQNpEv58jSJvoJzkb0ib6A/wK0ib6gxzB06v+CGdD2kR/lCNIm+hfMRE2UJ9g73yw9sqQprfTyJSDJQ+wd5z1K/giL9xW5WXMW9wbzcknL6aLVYdQbtkCjO6LKyRpFR1VdzPa+aB9g3C8E6tSGG0dhkr5NBqadFRdAdryWaDtJTwJ2vdR4/LzUPbXhP+SjurR9LAkX8jqOIuCugY6aqafQuHOYAXG2/GDB2YcJvwdrqPchVBllfIHmkRmPk8X7qAiR+iY+TQhPyTgJSCPEfIBAf8O5H5CKimCxwNL7ftc0C+QcWEPv9widXj0lkJ19MDcUid3m/AOKEAPXiG+SA9vrc7TeMl9DVZJpxi89yQBFJvVyf0UlpEU9jZxJpfxnlkzeztauRFtLWHd4N5nUD6JhpvdUFIPE62jBqyqFN1Q0graDYRfVym6oaQTJv8W4c9UCvPqrOwnhH9UaZnX4OY9Zpo3IMyLVXpyqbLFz827hh5UVtFRD/Mys15RZW66wsz6JQKO0VFrfC6z9tNTQzfsgnFQe84M27is7HAhw6aZYSe4YekZ5XJEWJtkNFp5heCGSCQNArb2IjgBGPfKOZeYjAY1es4tJgx+mtARPc2dS3ANC4XmvYvfn9NhLCie/zqdcYRul3UyizxPLw8q+gy9VC4rCpExPXOlYHA5kUjQdEnRLwguDZ4fXE9hvNegK1jkJQHzzWenIs6OeBjXQfOXy4oeDJolZSIEShyClpCMoC2Zbmh069WRpyuR5fXl4Ndnk+wm7PjR7ih4YTWYQsjZlHK6atPLnKwe6r/ysHxhtSAtqF5fjhZ2Vt2zshpVVKhO9spnVmIwkQjNRZvenstXGZQhC2KqauWW8m752ZXBvsr+4I7gdtbealVqEcw1tx2Sib2FyDMOU7DlqTq+Eq4myOCvJUQyeercWtcH25nCs9SszqhHeDc7LKHxCd28Qv6RRVBh3TmcpLGmMdJU3W53qTfbA1TTA3yio4tMH9Bye8mf20EBr7w8iHK2XwQXEtdMuMA03vMlSCLo5bLTewy394RKHP0EQqnwoDK355U7Ha1C+Nc04V/T4UyVbs+pyu6DLL+hrs3xkRpnX+bp5BMrq9G1tZM6ZF0eh5yVx/nq8zjpbPCtL2+vhvYNeV1yjsMl59ouOc/2vPkOZVFiQVguqOxCsxhv2CJnw1ycZwVlZMrObrerPydbcD3Tr9FpV3p8b89yvKb2rKvNhQZJS/Ygac0eH/CBttwR0p47Qs61uQoMl1yXV8So8JoTpGqPqDxzZfbA8k9nHYgNm7mLBoWLFtsuzQaIGAym5+vmiDTcfh9yj5lS50AoE+OlHP5fkTtWzQGiuqc/8q7KSZ24Ko8T5x0xGOYz8jrqTIej1tqOWucYVrOc460e/JgKC7hEQ7Zjzcn2kLnuGXRersvMz3UZ6TGsSpGSiCzkP+sl9L9s7vsqyU3j6VRTPDEwMj4Ya9q35LzIee2NI/HE+L7G4cR40+54Ji0xlpH47qbhgYFcliWitDS1qKbM/rFYOouRS08OREdYHd2WhO5IbF+m4GVJMn8ZH0ZwgX1144lhbMmbiQ/FY4PhPWhcOjM4GBsimKJzVrMgp8YHMpG16zeuIQLFIzaEX+knCSR7pfensWnvSDydMYuNRjN7zEsrUnEGksd1xxOZWCoRHYFsiu87do/HRzKN8YT5UwZ1keT3iEVnyiuSrCoV6iJfSK3VKtSqHWrtUlXxLVOrfLJavUM1dqjV29Uq0JrK1WpfaVgNbVCNcXXRy6phci4iTsV3STn9VZSXE6L6blSL2PkStaiiXK0heG5GrelXQx10YVhVlSCKKAlZnaV1qMW7iLVHDRJbseInFmVvh1m4qlydT9QlGXV+h7pkg3rOkFrTYUo5rpI2shZWFSVxjLiPs/JFvKASkMvVJUSoWaYuWU8FN1gFb1RDqKgf7Sgi9dUKUxWUvJjrvZHpPcunXE2cKzeocw+pNRmzwHGn9sVdVP3VL/mUaX+AnDlUuKpD1TuYCXQm6mLrrNScUmVf+aH6MEpf9RI1gFlZ79jg5PMd6yCTXq4q5XQqPaS2L0OvtNu9YlyiVm1jGBkAbfKtVkPKvg5GMqCC1TVVomsqVB/AE5ahtH4C76zk7f1HVcNZ20n4Q2o5U4oZR5E8Xg9rqurbYRYcIjAuCl5MBaH9o6xi9bkhoislQ2qxqRkT4VMG5IxPaehgOkCBMiipDatlSwk1taT+IMsVd3g9ZCPl3QP9qjJNDfhkL3ziHnmI6AHlvQMdh7wQodwtyzBmZSYcRo2j9epMCKEuXqbOXLpB9G0AaoDjYuXqoalYqGPLhiC1YkMHidWUbfX9HpUmpu+xqeqp/49T1RnNQGc8reWdqqxdQPCjIew53mpOKVmTkj1XsR3YzQlpPEHzz6AJD8cyybFMZCCZipllugf3J0aS0UE+YWVNa6o1ran2tKYUntYUe1rz5k5rbFsMwKIBZomVyeRILJowy5iTnXgzptwuy3xYqVXj6gKaJy4g91yuLiAfmNmvLuhQZ4bVUnI9n1bOnLXoCoJbMWrDaqARjrf/QdXvk/tVDd5zqbpU8XvOks2J6xhhe8MbqMTH5eVsPCg+z5wK7vWbMSYw6JazsUajItAPUQR9ibOcYMNG1jrV4rWq3FlfrxafhTrTz6nKLhfvPfBTtfhVVX7oOYyAkX4fiSNt2mTMn2yK8Yd5oQkx8nxom+fsaQQbvudMLMwwVb5WDfvegHnCZJ6lF6vhperSHeqifjVA8xE1v1+Vx9XAelVuginCvvgZ8zaavI0w+2y1sZHms34az7ry+wMYzlXKXfJLpEknxr7y1gHckKbRBFyJGTHcH/Yya91HfcfmMLI69QeTfrxeDcwi7V+lljT6PEsr81fzNq/mYUc1Py1QzWNTVDPPpzx6vWcp/KgSCN0ggl1quzLynNqCGbiPplpNZh2pvO7bMY0L26jqslrhkzs61KVPqEvA2W/xaWU0txLfBrUCtxZMicXKMLxKU6rlMDEpmWOYlcPqQqUYXXsnm7a4jZepDT7l36R+NL6BGq/MVhto+l/m+xiT2Y2qMqsfy/rlG5BzZx/y+Junsezp6f91+PW3RFx8FuARF00pfBoT85XHMY2ZwVQ0NWzC2XFRwWDMa81aXnvW8hSetTyTBmNroK+YAREgmiUk9SyaqESyTymhzvf0l7GpBHc+n3Jagg8rSqd5Jzwt9febuGzh/RlO8WnLmI93yPwm2m/69kqKiVqUc7Q/UNx0yLPGFM8ddIdZjS4zGSmtg4nVzUpTGZ/2KyYy4BKZIZHVFCPIPs98HWcKDxnaz8ky/u/Hf4fYvTQSSQ4NRfBJi5Wp6MCVsUyr+bkKKbJ+SyQViw5GxjIpKTKwJxpP8HA6lsDHPKLr2YcwulKxiXhyPM1LdUdWjUTTNraSSmyOjsbSY9GBmBRJ76Hoe/f4EF0ZTGYim2KZPclBJq2HGZ/AFelVe6KkCvWSQxMqw7a5IobuGPu2Sm/ySlIkEkmnBoi4OpUc4+yFPpbCaiTHiydig9JIMjFcz/7Di1y2PVQiabc5loBWq+OjQoWhkegwPugBmS61wBsZGE/RTXdkfDTB6KhxYKLg91YiDl+VIsmRQXQCXvV3R9aNJHdHR9YkJqRoKhXd7/ySSzwxwky/Ojm+eyS2KknqW3buGYsNWMjG2ERsxO6DntRAKgaT96biY9kFTbWgESk9Gk1dGUvxXcgiQ6lYLBVLR8zegnPELImbo5tNC44n0vFhsqdpRZLBds9iLkOeMJ5KUUizZt8YyUnjOzFUMja8PjEkMT/ILo3dzlqhZHrMYeB0dCImOn5ztDsWZU41Ek1caX6ShrtNJMLmjngiwj9lg2/TXIm2Uz0bycpZ/dgqRUZNl7Bckwye5bPoALI0+bbVbNbgQWZ7qSkzOtbUTf9t2NO7dnBHU3d4/eae3hUbN7ZFdy9uazu3uXUgen6T6P0m00d7YrHBtKNtpq1MV4Nvkh5XRofZcCHK+sTKVfjgToqsl8FHd6gEGhUxDY0p0DKFGDtdyTSsO0QBX0bqWdPfxSWviyViqfiAxLs4bRpTOLfVDTTIU7xrrP6MD8Yig9FMlExset+a0bHMfpiHOXECu8OZJhMtm8iwXee4R6/b3Fe/qmWx+Yv1+sbRzHgitmzY1IfQaGpgzzK6ETWe117fOMz+bWmtbxwaw9Uh9pmgRu7/yVQjzdBJ1lDrU0JmF4quc37qSHwYKbKJujrL3cPojC7T2PwCnI95y3gqy825d3Szr/2kyK3QLBrHGXuq4M4JM9O0RPG37bdm124dJ4XssRlNZyagujAXdfvguaIyQRxPjKdjg+SnNGOkM+ZU5GRAhw1HR9lkMIJh4JiSqGHjY6Za1iBmE013pC+xOzmeGDRNR002e7Q7udcp2/UpKdtTeUv5fYJzw8u616ximmTsOWdTnEZ8YnhFalgamHqbQMe3noZT0dFIPDHBPy1l3jEiCYwJerChf+Y3pSKrR0bI1ZO536JybeTX2923RkqMj5ImaUiAv9LMMWLeetJZn4tyf+cqjzpMkXw7JMIzaG7hH7vKt68iVR4nh5hsg0L+wSrrUTD7q1OrN0bW9m1eJe3dEx/YA3HkbihitWY1zfLwp8ye3G0Oc765Jb7uRYqPxRLU5fFoWlq7YmPPGtg0MRGnoYZS6fjVsUjG/IhWgQ9aZX32K8+3sZKp+DBpHXXcD6lNydi+GPkG+7zWIN2+qbKh8cRAJDL5Dpb5PrRFCuAl6vCYmHWsb2hZ4aj1eS37VjERNccFBcDkXBOxFM2beyNkVDpiUQej+1NdTBDui5EIhc0DliD3B76GRjOSVdIsxW7/ZCBzgmD1oNHEHB2Thiz9WTwCV4vCbSOiBpphhqXsX6f+VCxblvoE6LE+fdkjQLHxhlf6WO47VWhvlde5gDctmV0CfNmSeZnYqmarYpG2KdnVyGyryLx/v+MC/82uQ4CfWXXsEnX0WQJ35daxq/AuMR9wkX+xRO4UoKilSDogtc5rqzsl6uqxBO/scVVv8eYqsbOwEsL4PkuJLgG+aneRqH6zJbJHgNssUlduxT35Ky3ikh+1KxU1XJgr7lKblG9PJdk8sPyPi1vtEdsTNZ8ScL9VVbcAd3rErkYp+dQ1HPmyzSfAB518uXp2u5uNaguZezIVbWtsy61l2+epBX+tXFqfJbV31NUgIomK/FZFvQIssUj9szi4yOYSYKPNlV8PMSh/bPe1AP9sj1cBSpP19a850/vWdlETkvkT7ryWEOy/tf1YgO85LFCw3HueyfzO8TfN2t6rT4BdFqmnyzUZTja9iQ286q3iXQK805qLLrvP9XnkgjNZs/AFW5wAj1tlt+WK25Z/4zK3uBJpzPNJY9snjTM+kVrnnnKLLpEmnJfd1bguFzKuMF+v3QYBHrMU7hHgffaMndusnvwbsvVx5EBu2cds0lM54nK8wufhf1lid0v2gBXwQxat5ZSAP7RoracEPFe2ywr4gGyXFXo+Jdtl3YqCz62phmkfmmbfBCyd9zp0FvAJh84C/p7t+QKstuyzQ4AzZTG2xz2nBDLX4tsqwAVOPoGss4nXnxJIjx1QCHDPJGYhXQT4LbtkjlEUe+evyTrPaYjpXNzZltidZ+foeEk+HXP96ZKptpVzew+ZToCvW6RLhO98YpH6Bfhni7RLgJ9JYmvAa6WrpAaBlshiB8D9fIq9VmoQ/XSl3XW5rdiae7MXn3SdzLv+3m7T93LbtO5vqtmaBvm2F2u9Vt8L8BKL1CPAXTaXAMW+GYWn7hB3nZnW/atPgHedUVh7LudabN8ABbjUlinAbjsu2O6qJs/sbf8t50yr7FpW5UTKPVtzQrteUcugXXAwp+Ke3OlVNGu5ZYTACs+p/LeXz2eC5ZO39yCf0m63buJ9AvSeUYd8lXM/ZkcBAnzclinA79qx2w9c1UzaIU9w5qftWgT4HTtsEeCLdtjyg5wgvevVnIrzdIholqh4sg75fCZ4wjNFe7s4w3a7rACvtssK8HrPVL0k5F1uW+XyPIV7C3RxPedZYKsjwA22OgLcMqU6Ql6zrU5znsKF1DnIz0fsW+oRxyMO71ABVnrOxIezJf/QktwrQI8d8gnQLzt1FViFxdgrwIvtsgLcnFV2s0vXSQZCropbBejUZ2sefbZW5DHJ1lOFtt4Vf4LhJ9Zzh/+zUwL5ZzvWFKDXjrUFqNmhhN9lACZNIL12xCjAX9rSBFjmcRQtZKe3+PkdW8E6V/BEJPGo8JBNErW8JZ/RM03QdkIBil8In4nHOQX0CrDDIm0T4CrJ6S+rXDUV8BfxIeflFttluSULpnbEQ8oS24ICvMB+ZhMCN9pcG3PqyJldMe8+xCffJx2P6SoPH4q8BYt6JDdTvrBdls5AlPn3Gq//V/bNVIAeq+xlQkyZd0rD2U162hb5Igd/kVvLuzbp05yKu3IrLuiPPxPOYg83AT5pkXYJ8J9sLgG+LE+ZjMutoze3jsuezHmG6P1WTh2XnXkdW3Pr2Jlbx9bcOnaeeR3duXVsy62jO7eObZPX8J79dCHAF+zb1As54j5H9+5wq4zpYcepXK373Vor0vH/uC7fzH8TZ7jNrkSAh61Kjh88kGfE5RbtL8xkCwtIoVNT8wUk60dJk06pd9vzsQBfkc5gyErSHfZEfEdOyZ7JSh6278WHc0puO7OQgzxLgE/YDvOESxz6tzuPtcQz4TcsyvapmQqZ3s03melFBvBtVwZRJB7dfyc4/RFP9rPDEwX45/EJsVXJfvhbrkw+84oJVHWdQ/xc4T2z8lPxj/HzBD9f4+rWm1xu6bat+08M6O38HOHnIX52D4vP86dQ3y7x5KN7rCeDbLpife8gm+61vouQTVet7wtk033Wdwmy6UXW9w2y6Zr1HYRsut/6/kM2PSDV56UHpZ+V5qMXS/+Sl15ifU8im67beeYsuiE91JCPHpLq5+Sjl1rfp8iml1nfpciml0udeekVeb+boEjTpDcuzEefbn2HIpteaX2XIJteJUkr8tGrC/hVTQH6jAL0mQXotQXodQXoswrQ6/Omr73SR5+56cUsfsz1B53TQ2XZ9Nmc3uWiL2Z12HYTw3Mtg3P7d5TLOeaSs5/x5/rJvQX0f5jx65LYcbvTMc/IDr8V89HTBeR8m/OLTX0Mx2Mm9HRvq/XPnF9swVTqCHZAF7ubzXSkWJk+Ljq2/ZAdfv4Gp5fK+fWskU07ixuSmA8K9e8cxh+Squbn+kk+/hYZ7c2dBxYzObnjZTWj547HTbJpRff8cDmX86FLzgij547TQ0yf3HnvVjnvVCjdgV2tqL/cW6h/rQD/UwXo/122Hx+df78swP9bOe8UKf21AD+e4fPJn1eAvsSTX86KAvTeAnLGCtAPFaDf6THt6d6u4esF+L8l+F27Yr5cQM83OL97S7/fF+D/tABdVkw57l0Xa5X8/XIO53dv3tep5G8XllDkq/cSLse9edfeAvzXc373bpz3F6j32QJyvleA/zUu371b2/8uIOcDzu/eB+Ivol2u3eMRB+azZ9ibX/6F3vx6YmP/hnw3r4FUJp0ZHxoKD0iDMbH0KpIZxQorLL6KRAaTkWG2qjgymEmm0pHo+D5pIDk6NhLLxAbDi9taz83PFBmKJ+IRtgw5Ektg+eNQKjoaiwyOj45iKaQDM9epOVmtpZbWmnFrrbG1vE0s+Aqf37ykPd+SNevy+W0OrKV5sRNracta0Get+HKs0lrbvWLTmsiazasjEYm1Cg1OJyN7oonBkZgUWb1j84pN61cRdd3mvsiai3iBi1ZjeVzvplWi6LqNW1au2BjZsnZtz5reSO+KlRvXEBVtF8tNCyz4dq5YXbE9Z9FiZ+e6jetXroq0hlvD5+YuwOzdFLG7ltRBx/ZiYVzexXBsRV22RLYUMqeSguv13GsF7ZLtbPVltiSiZBPyLgtk6yJdKmStbnStLcy3xjJ7fa+9ZrTgespIZHgUa0jxewHnukLHqkxzvbitWEu7WIU56SpEsXpyslWtrmWDdh1tZEfHun4srcy2jGsNq2ONavYSwwI/dWAek99f3EtcxULJ7PojkYF9UQyU6Ej8alevSeH0/tFMdDedMynzvEdACWpQeDgxHsbKycHG+KDEsD3R9B4pPLg/QSXNcyZlXpmIpdgPA5xIhK6lYiNRMHJobCQjhdkwAxgeThKQwe+Qwmwwh1NJtkQ9HNvDJ6U9gykbM4uas5NZQsAohRpIANMLw0iCeLNGU+judFoK03w5Ci8ND8Z2j2OBaDQxTCOFo3GsRBaXdu9OxSYExtaWcpi1mpOTAwLkkv6rfwg3iiT7u4HiuV2c5+bEsq77ER3/57PPkqL85a43gF2ORDx7Q+OOt/ibCVFe5AHE+UPFrld2lBffI+zkskV5kS8Q50OeyV864bM5nzn0F3kFce516e9xnbGA4q+O8iL/IM5dUn79xd8gvybKizyFOHd68ttPtD/By68U14uzz2OO8tPzlN/H9bLSLKHsc+UU/Z92lRf5EXF+0mVwl3jpgKu8yKOIs9temut8i6u8eL4W5/WB/PWLv6Ou8uJ5TJz9U7T/Tl7ecrP67PNDLv9zx3T3u8qL5zhxvnaK+v/OVV7kg8T5Pz357Sc5ntN1h3+JfEBz45nZ/3n+DC/Ki/zDkjMs/yOuvygvns87G7Of0wvZ7w3ed6K8eD6uCmfH015XvcIPDrnqF3mtN/iDysIp9P9XV3nruZqvzHx3ivLvu8qL5/53mye3n/j7kNOs+nm+4sPm7HFYyH5/4vU3u+U2Z+dP3POnOP9Vyv+N3eIW83xsipf+JXL+8id5+S/Mn7x8dYHy9fyV0PlFk5dfUKD8jgjPJwUnL98m57df15X85ad/cvstL1C/Z9w8/8+qyetfX6D86aPc/vLk5bPmLqf+/NXUaXE/xDNmnvuH3+l7znn2y+b5Ku/k9ZcWKH/DA9nvJwqV/79QSwMECgACAAAAPUnOUJMG1zIBAAAAAQAAAA8AAABmaWxlMjJhZGU4YzUxMjYKUEsDBAoAAgAAADxJzlDuLBUEZQEAAGUBAAAfAAAAbGlibG9jXzE5MV81ZDYxYWUxNDIyMDgyZDZlLnJkcx+LCAAAAAAAAAONUVtPwjAUroMZ2ZOXPzDfMNbBRIw8aCK3hISLcWj0aSmjYw29zLZT8c/5z9QSBzH4oE1Ov3P5Ts752gcHAFAAhW1jtnGBfTfunlwAYB2YwDK2a2wLFEHJ4HtFs7Rya65H+hY/NSsaszSkZJIinfgNv+7HCNfP4ga8x1IRweGNJEISvYBtnGI+VbDHUiG1gn3C54TPxgIG2WyGlUl1eIJ4hJe1CHOFVxgSFXZHQbCOpaFLEmkVZoY1CkK9SDG8llGi4KBdVxmDQ4ynqiVYSijSy0WaGaE6l7OXy9lhhJOpeOF5XP6vvJxvV72q5xun9GnOGr+Lh4lmVAtBFXTmnGgJHcmQnC/nQec1zvhGw4E2snSCtFu+unRPPd+rHm1Q9ge9sXvsxoRit99rdYZB59f0vzBfveadezUAitaP77U5Yljlj2TlyeIEKbwiPCOaYeN8fAGtIecSOAIAAFBLAwQKAAIAAAA9Sc5QGWZELJ0EAACdBAAAHwAAAGxpYmxvY18xNzVfNTY3ZTdkZDExMjVjOWYxZS5yZHMfiwgAAAAAAAAD7Vnrbts2FFaTKPfOGYoB7YauGrCh9SDLli+p/2xAl6RbsaQN4nbYv0CVGZuoLgZJpQmwd9qT7RW6HVk6FCXLTtwF6TDYgEzynO9cSYqH9u+bmqYta8ur8OjQ1fQ3r5/Xupq2dA8GS/DswHNHW9E2oP2zHnFW90LX8eoefVs/ib+Zwy7N3wjjNAzMY0ZDRsWluU9GJOhz84U/Cpng5iEN3tFg8Do0e9FgQDiQDoKhE7gk5rkk4ATbU8pPn7/q9eSYAZxRV/DTCFCveqfickTMZ8wdcvNov8Mj33xJSJ/vhf6Ieo6IHfkpop6ACMD9O3+l7q+8dTiR/TAUaV93PYfzdLDmehEXhKXDDTfsExGGHvLX3bEVCVjvOwLUCil/FjJCBwGyB8wZDamL7I0B2yfnEBgSVgaM9tP+5q+EBT0fPBuiNghHABrBR896vbS/euRAVi4Q6BMxDPtSqz9wz7EfeL5UEAQEw14fOczxPOJhGhgQkLnGR5BIx8uGHg2k0zoXjox4dTxoo1YesXN6LiV14XrinRwomdQjQeXg/rSlteAv+Av+J+frLWvXauGOt61WrYk7fvUpjOwOApuWbTWQ1bCaNXu3TEdeYcPq1p52rsTlBmtNq9nKtK+BrUat1cX3UOxVx7Y6mc/gSndKON1ay0Y9Lcuu2e18dM1SuXaM7OSA9nwx6LGxUqemD9RjbIsROI98OGpJ/9+S8sfjTBvToLfh2i0q+cgwZ4tBDfA3fFLawxPjyY8/GC2rYTWqpoHVgmmoJ+yXedCYZRrq8bmNCMBUsZpIaLAfq9Jo3vi2qncCk8feQ22dzIdSNQmtjjQ7jSuteiZCnBJHU/Uo4T2S8bTHOuOyBo3UdtvVoop25s1XBdGc5byVqoR2ymZECSUV+HbmDBaNXDER7VkTUS6r1oTfZNaVjM/hyBdpGZmi8/Os1K55qR3FWFlOH8c17vUXQaGWtaRPanSxQlXWNPKl8nZJAA2pSc4lKkyKaWAkdW7eoUfl7mYm8+FWxlWuacibQnmuK1M2e6q7JFt4VmeC+YAwzMmwMhzGV5J3ffYK+b+3pYvvs3h3QbbwUlPYtwk3vYnNZ2iuy97XGRjMjY2m4IMLwRw8TFJa8m6Eb7jsKb6ndsnFyE+U4L4u3nHv480wE84JbP3iU+7mSKqzhRQWb4QqtFJYoCn5iZpX0+hFo9E+jX8zKCz49hU5/l5Gpkhe+F4TNmfEvPEW9cPAdxjeUB8UYRngdtaf7g5ZGBSI3yWZOIJoPHwfQFpgmjg5glyc8XhCrmmhwoPwPSyO96DpBN5QNxxA5dhhwgjPjBNDrVo33gQe9amQVdLdn48Pa03jDyNuEbUJg+T8xaN7GSizNU8hT6qaT/66cUzaKY3sQQYDzhmcC8bhi72Dl3sH84ZxpaZS+1PUl2LnUTAfeevwpmflU58ai3bRLtpFi23ymloKQjy/LuVNUenG/HLJcvx/vPvxIVz/d7LF4IYG2kr8vx7+p7fcpz40d7Xk776H8HwOz4cCTA8cfzzRO2NgWskrP27pUOBH8U34wz8K3uObVBwAAFBLAQIeAxQAAgAIAD1JzlAgd+Ci0DAAAAQxAAAfAAAAAAAAAAAAAACkgQAAAABsaWJsb2NfMTgwX2EzOGY3ZDU4MTMzNDRhZjkucmRzUEsBAh4DFAACAAgAOUnOUFsx5RUpPQAA2KcAABIAAAAAAAAAAAAAAO2BDTEAAGZpbGUyMmFkZWNhOGQ2Ny5zb1BLAQIeAwoAAgAAAD1JzlCTBtcyAQAAAAEAAAAPAAAAAAAAAAEAAACkgWZuAABmaWxlMjJhZGU4YzUxMjZQSwECHgMKAAIAAAA8Sc5Q7iwVBGUBAABlAQAAHwAAAAAAAAAAAAAApIGUbgAAbGlibG9jXzE5MV81ZDYxYWUxNDIyMDgyZDZlLnJkc1BLAQIeAwoAAgAAAD1JzlAZZkQsnQQAAJ0EAAAfAAAAAAAAAAAAAACkgTZwAABsaWJsb2NfMTc1XzU2N2U3ZGQxMTI1YzlmMWUucmRzUEsFBgAAAAAFAAUAZAEAABB1AAAAAA==" download="RtmppcWR8N.zip" class="button">Download Rmd</a>
#> </aside>